fixed asserts due to missing MOAILuaState GetValue implementations (s64. u64); changed MOAISerializerBase and derived classes to no longer use uintptr (now uses void* and lightuserdata)

+++ src/moai-util/MOAIStream.cpp
@@ -265,7 +265,8 @@ int MOAIStream::_write ( lua_State* L ) {
 	size_t len;
 	cc8* str = lua_tolstring ( state, 2, &len );
 	
-	size_t writeLen = state.GetValue < size_t >( 3, len );
+	// TODO: 64bit
+	size_t writeLen = state.GetValue < u32 >( 3, len );
 	if ( len < writeLen ) {
 		writeLen = len;
 	}
